+++ cfme/test_framework/appliance/env.py
@@ -1,0 +1,3 @@
+
+appliances = load_appliances_from_config(conf.env)
+reporter.write_line('Retrieved these appliances from the conf.env', red=True)
