+++ cfme/test_framework/appliance/regular.py
@@ -1,6 +1,5 @@
 import attr
 import pytest
-import sys
 
 from cfme.fixtures import terminalreporter
 
@@ -9,26 +8,32 @@ APP_TYPE = "regular"
 
 
 def pytest_addoption(parser):
-    group = parser.getgroup("cfme")
+    group = parser.getgroup("appliances")
 
     # regular appliance type
-    group.addoption('--no-regular-apps', action='store_false', default=True)
     group.addoption('--num-regular-apps', action=1, type=int)
 
 
 @pytest.hookimpl
 def pytest_configure(config):
-    if config.getoption('--help'):
-        return
-    reporter = terminalreporter.reporter()
+    app_types = config.getoption('--use-apps')
+    if APP_TYPE in app_types:
+        reporter = terminalreporter.reporter()
 
-    if config.getoption('--use-dummy-apps'):
-        plugin = DummyAppliancePlugin()
+        plugin = RegularAppliancePlugin()
         config.pluginmanager.register(plugin, PLUGIN_KEY)
+
+        from cfme.test_framework.sprout.plugin import mangle_in_sprout_appliances
+        mangle_in_sprout_appliances(config)
+        # TODO : handle direct sprout pass on?
+        appliances = appliances_from_cli(config.option.appliances, None)
+
         appliances = [
             DummyAppliance.from_config(config) for _ in range(config.getoption('--num-dummy-apps'))
         ]
         reporter.write_line('Retrieved Dummy Appliances', red=True)
+        reporter.write_line('Retrieved these appliances from the --sprout-* parameters', red=True)
+
         reporter.write_line('Retrieved these appliances from the --sprout-* parameters', red=True)
 
         holder = config.pluginmanager.getplugin('appliance-holder')
@@ -48,5 +53,5 @@ def pytest_unconfigure(config):
 
 
 @attr.s(cmp=False)
-class DummyAppliancePlugin(object):
+class RegularAppliancePlugin(object):
     pass
